backend/api/requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
anyio==3.7.1
async-timeout==4.0.3
beautifulsoup4==4.12.3
blinker==1.8.2
Brotli==1.1.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
commonmark==0.9.1
dacite==1.8.1
Deprecated==1.2.14
distro==1.9.0
dnspython==2.7.0
exceptiongroup==1.2.2
fastapi==0.115.5
ffmpy==0.4.0
filelock==3.16.1
Flask==3.0.3
Flask-Cors==5.0.0
fsspec==2024.10.0
gradio==5.6.0
gradio_client==1.4.3
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.10
itsdangerous==2.2.0
jaconv==0.4.0
Jinja2==3.1.4
jiter==0.7.1
joblib==1.4.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mutagen==1.47.0
numpy==2.1.2
openai==1.54.4
orjson==3.10.11
packaging==24.2
pandas==2.2.3
pillow==11.0.0
platformdirs==4.3.6
pycryptodomex==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
pykakasi==2.3.0
pymongo==4.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-env==1.0.0
python-multipart==0.0.12
python-slugify==8.0.4
pytube==15.0.0
pytz==2024.2
PyYAML==6.0.2
RapidFuzz==3.10.0
redis==5.1.1
requests==2.32.3
rich==13.9.2
ruff==0.7.4
safehttpx==0.1.1
scikit-learn==1.5.2
scipy==1.14.1
semantic-version==2.10.0
sentry-sdk==1.45.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soundcloud-v2==1.6.0
soupsieve==2.6
spotdl==4.2.8
spotify_dl==8.9.0
spotipy==2.24.0
starlette==0.41.2
syncedlyrics==1.0.1
text-unidecode==1.3
threadpoolctl==3.5.0
tomlkit==0.12.0
tqdm==4.67.0
typer==0.13.0
typing_extensions==4.12.2
tzdata==2024.2
Unidecode==1.3.8
urllib3==1.26.20
uvicorn==0.23.2
websockets==12.0
Werkzeug==3.0.6
wrapt==1.16.0
yt-dlp==2024.10.7
ytmusicapi==1.8.2
